requirements.txt
et_xmlfile==2.0.0
numpy==2.4.3
openpyxl==3.1.5
pandas==3.0.1
psycopg2-binary
python-dateutil==2.9.0.post0
python-dotenv
pyxlsb==1.0.10
six==1.17.0
